+++ src/sentry/integrations/slack/message_builder/event.py
@@ -1,73 +1,25 @@
-from typing import Mapping
-
 from sentry.features.helpers import any_organization_has_feature
 from sentry.integrations.slack.message_builder import SlackBody
-from sentry.integrations.slack.message_builder.base.block import BlockSlackMessageBuilder
+from sentry.integrations.slack.message_builder.help import SlackHelpMessageBuilder
 from sentry.models import Integration
 
 EVENT_MESSAGE = (
     "Want to learn more about configuring alerts in Sentry? Check out our documentation."
 )
-HEADER_MESSAGE = "Here are the commands you can use. Commands not working? Re-install the app!"
-DM_COMMAND_HEADER = "*Direct Message Commands:*"
-CHANNEL_COMMANDS_HEADER = "*Channel Commands:*"
-CONTACT_HEADER = "*Contact:*"
-GENERAL_MESSAGE = "Just want to learn more about Sentry? Check out our <https://docs.sentry.io/product/alerts-notifications/alerts/|documentation>."
-
-DM_COMMANDS = {
-    "link": "Link your Slack identity to your Sentry account to receive notifications. You'll also be able to perform actions in Sentry through Slack.",
-    "unlink": "Unlink your Slack identity from your Sentry account.",
-    "help": "View this list of commands.",
-}
-CHANNEL_COMMANDS = {
-    "link team": "Get your Sentry team's issue alert notifications in the channel this command is typed in."
-}
-CONTACT_MESSAGE = "Let us know if you have feedback: ecosystem-feedback@sentry.io"
 
 
-def list_commands(commands: Mapping[str, str]) -> str:
-    return "\n".join(
-        (
-            f"• `/sentry {command}`: {description}"
-            for command, description in sorted(tuple(commands.items()))
-        )
-    )
-
-
-DM_COMMANDS_MESSAGE = list_commands(DM_COMMANDS)
-CHANNEL_COMMANDS_MESSAGE = list_commands(CHANNEL_COMMANDS)
-
-
-class SlackEventMessageBuilder(BlockSlackMessageBuilder):
+class SlackEventMessageBuilder(SlackHelpMessageBuilder):
     def __init__(self, integration: Integration) -> None:
         super().__init__()
         self.integration = integration
 
     def build(self) -> SlackBody:
-        action_block = self.get_action_block(
-            [
-                (
-                    "Sentry Docs",
-                    "https://docs.sentry.io/product/alerts-notifications/alerts/",
-                    "sentry_docs_link_clicked",
-                )
-            ]
-        )
         if not any_organization_has_feature(
             "organizations:notification-platform", self.integration.organizations.all()
         ):
             return self._build_blocks(
                 self.get_markdown_block(EVENT_MESSAGE),
-                action_block,
+                self.get_docs_block(),
             )
-        return self._build_blocks(
-            self.get_markdown_block(HEADER_MESSAGE),
-            self.get_markdown_block(DM_COMMAND_HEADER),
-            self.get_markdown_block(DM_COMMANDS_MESSAGE),
-            self.get_markdown_block(CHANNEL_COMMANDS_HEADER),
-            self.get_markdown_block(CHANNEL_COMMANDS_MESSAGE),
-            self.get_markdown_block(CONTACT_HEADER),
-            self.get_markdown_block(CONTACT_MESSAGE),
-            self.get_divider(),
-            self.get_markdown_block(GENERAL_MESSAGE),
-        )
+
+        return super().build()
